+++ changelog.md
@@ -21,6 +21,7 @@ Codion Change Log
 - DefaultFrameBuilder bug fixed, now always makes sure the frame fits the screen.
 - DefaultComponentDialogBuilder, DefaultListSelectionDialogBuilder and DefaultComboBoxSelectionDialogBuilder bug fixed, now always make sure the dialog fits the screen.
 - DefaultFilterTableCellRenderer bug fixed, no longer sets tooltips for null column values.
+- TextComponents.preferredTextFieldHeight() removed.
 ### is.codion.plugin.flatlaf
 - UIScaler bug fixed, now supports the codion intellij themes.
 ### is.codion.plugin.flatlaf.themes